Refactor: Don't format the log message twice
Previously we used a separate log object to send data to standard
error (which is to be consumed by ns_server). This means that
the underlying methods (formatting the timestamps, adding the
log level string etc) had to be done for both the file logger
and the standard error.

Later on we wan to be able to use the full features by the
spdlogger (and don't go through the servers C API). We
don't want to have to send to both the loggers, so this
is a preparation for that.

Change-Id: Ia463770b9277602ad875b12bbc138cb2b53873e5
Reviewed-on: http://review.couchbase.org/88234
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>