MB-27063 Fix AuditFilteringTest test
Fix test that incorrectly waits for the string "uuid" to appear in the
audit log.  It should be waiting for an event generated by the user
"another".

Also increase the maximum wait time for the string to appear in the
audit log from 5 seconds to 10 seconds.  This ensures the test passes on
slow disk systems where it takes a long time for data to appear in the
audit log.

Change-Id: Ifb80f1a29ac51d620052690c42405b20490b6eb4
Reviewed-on: http://review.couchbase.org/88235
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>