MB-27457: [5/n] Allow KV-engine to set the delete time
1. Update queueDirty so that it assigns a time to all deletions that
don't already have a time assigned (allowing DCP to see the delete
time, not just couchstore).

2. Update couch-kvstore so that it no longer assigns the timestamp
to deletes.

3. Update DCP consumer so that a V2 delete interface is available
enabling the consumer to accept a deleteTime and set it into the item.

4. For DeleteWithMeta, the existing protocol 'expiry' field can now be
used to set a delete_time. The old deleteWithMeta packet though just
copied the setWithMeta meta fields, giving it a timestamp field named
exptime, for clarity I've changed the protocol struct just so we have
a more distinct name for the field. With this change a client can
set the tombstone time by passing a deleteWithMeta with a non-zero
delete_time field.

Tests are added covering the DCP consumer API and DeleteWithMeta
showing that both interfaces can set the time when creating a delete.

Change-Id: Id79ec006ee37a62d740827054b8b369933dc3253
Reviewed-on: http://review.couchbase.org/88301
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>