MB-27700 CouchRocks: Block Cache size as ratio of the Bucket Quota
The Block Cache is the read cache in RocksDB. The Block Cache stores
data, index and filters blocks in memeory.
So far we have exposed to the user configuration the absolute value for
the size of the Block Cache. But, what we actually want is to size the
RocksDB internal caches with a baseline value that depends on the Bucket
Quota. So, with this change we introduce the new
'rocksdb_block_cache_ratio' parameter.
Recent Perftests and further investigation have shown that the index and
filter hit ratio is greater than 99% when the Block Cache size is at
least 10% of the Bucket Quota. Thus, we use 0.1 as default value for the
new parameter.

Change-Id: I16be9366cdc408592d2cbab9b8c05ac21104f077
Reviewed-on: http://review.couchbase.org/88293
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>