MB-27457: [6/n] Allow DCP producer to send the delete time
The DCP producer will need to send any delete_time it gets from
disk to clients which have enabled the feature.

1. Add a new DCP open flag to the protocol
  1.1 This flag is not yet known to the mcbp validator, so this WIP
      feature is disabled to real clients.
2. Have the producer track the state of this flag (on/off)
3. When on, an alternative dcp_delete callback is used which can
 accept the delete_time parameter.

This commit only updates KV-engine, subsequent commits will link up
to the real connection.

A test is added using the DcpProducer to show that we can retrieve
the delete-time of backfilled items, in-memory deletes and expired
items. When an expired item is produced (as a dcp-delete) the time
it has is the logical time it expired, i.e. its expiry-time, not
the actual time ep-engine deleted the item.

Change-Id: I76ddbbdbd33f14fcbabde4eb77d29713a1a9fba6
Reviewed-on: http://review.couchbase.org/88302
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>