MB-27699 CouchRocks: Total Memtables size as ratio of the Bucket Quota
The Memtable is the write cache in RocksDB.
So far we have exposed to the user configuration the absolute values for
the size of the single Memtable for the 'default' and 'seqno' CFs.
But, what we actually want is to size the total Memtable Cache as a
ratio of the Bucket Quota, and then let kv-engine to decide how to split
the total Quota among Column Families.
So, with this change we introduce the new 'rocksdb_memtables_ratio'
parameter.
Recent Perftests and further investigation have shown that the total
Memtables Quota has an impact on the throughput (MB-27552). Setting the
Memtables Quota to 10% of the Bucket Quota seems a reasonable baseline
choice. Thus, we use 0.1 as default value for the new parameter.

Change-Id: I438f9372d20c8ecb2cf51d140ce5ae4cd4885be1
Reviewed-on: http://review.couchbase.org/88360
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>