MB-27348 CouchRocks: Expose RocksDB Block Cache hit ratio stats
Beyond the raw Block Cache hit/miss value, we are more interested in the
Block Cache hit ratio for the data/index/filter blocks.
With this change we expose the new following metrics to 'cbstats':
  'all' set:
    - ep_rocksdb_block_cache_data_hit_ratio
    - ep_rocksdb_block_cache_index_hit_ratio
    - ep_rocksdb_block_cache_filter_hit_ratio
  'kvstore' set:
    - rocksdb_block_cache_data_hit_ratio
    - rocksdb_block_cache_index_hit_ratio
    - rocksdb_block_cache_filter_hit_ratio
We also remove the previous raw hit/miss stats from the 'all' set to
limit the size of stats on disk.

Change-Id: I7fa1e5b349ab5767af2c0305b6121cfe934bb0d6
Reviewed-on: http://review.couchbase.org/88421
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>