MB-27457: [7/n] Update the dcp_deletion protocol and executors/validators
Update the protocol so that the update to DCP deletion has its own
clear packet definition (v2), update the validators and executor
callbacks so that we can send and receive the new packet definitions.

Note: Externally no client can yet enable 'delete-time' as the
dcp_open command is still stripping that flag from the open command.

Change-Id: I69cdbdafb3b2bcf37e9e7f76ab9ca51b26c0bba0
Reviewed-on: http://review.couchbase.org/88428
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>