MB-27701: Use a UUID in the log messages
Generate a UUID whenever we build a connection object and
allow the user to reset this to another value as part of
hello.

If the key (agent name) starts with '{' the string is
treated as JSON and we try to parse it and pick out the
following attributes:

   a - The agent name. The server will log the entire
       string when receiving the attribute, but
       only preserve the first 32 bytes and report
       them as part of the connection stat group.
   i - The identifier to use in the log messages when
       referred to this connection. The server will
       log the entire string when receiving the
       attribute, but only preserve 33 bytes and
       report them as part of stats connection
       and slow execution reports.

The log messages for slow operations is changed to
the following format:

  <id>: Slow operation. [json payload]

The JSON payload contains the following attibutes:

    cid - The connection connection id followed by
          /<opaque> (the opaque is printed in host
          local byte order)
    duration - The duration for the command
    trace - The trace information for the command
    command - The name of the command

Example:

    35: Slow operation. {"cid":"c21fee83af4e7943/c21fee83af4e7943/0","duration":"20 ms","trace":"get.stats=331221786765:20721","command":"STAT key: dcp","peer":"127.0.0.1:51037"}

Note: Clients who don't set a connection identifier
gets one assigned to them so that the log entry
won't be empty. The assigned id is the peername
for the connection (ex: "127.0.0.1:12345")

Change-Id: I47fa596013ae3156895e01a5e461a0600eb7635b
Reviewed-on: http://review.couchbase.org/88475
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>