Replace blackhole logger with spdlogs null sink
Add a new property "console" to the logger config
which may be set to false (it is enabled by default)
If console is set to false we'll use null sink instead
of stderr.

As an extra bonus we'll use the colored sinks when
the unit_test option is set to true.

Change-Id: I690b7022e2cef04b98fd8ebc4ddd39d9913a0d42
Reviewed-on: http://review.couchbase.org/88493
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>