MB-27755 CouchRocks: Set default options for RocksDB internal stats
RocksDB Statistics provides cumulative stats over time. With this change
we set a baseline configuration so that we automatically dump stats to
log files every 60 seconds. Also, we set the stats level to
'kExceptTimeForMutex', which means that we collect all stats expect the
counters requiring to get the time under a lock on a mutex. From RocksDB
documentation:
    "If the timing function is slow, it can reduce write throughput
     significantly."

Change-Id: Ia9d56d5b97500969029b56163cd08700eff1b05b
Reviewed-on: http://review.couchbase.org/88498
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>