MB-27669: Fetch uncompressed document in 'passive' mode
In 'passive' mode, documents are stored as provided by the client. On a restart
(thus resulting in a warmup), we don't maintain any state of whether this was
initially sent in as 'compressed' or not. If we went ahead and retrieved as
compressed, an uncompressed document, that could be KV-engine actively
compressing the document which seems more of an 'active' mode characteristic.
Hence, in 'passive' mode, store it as uncompressed.

Change-Id: Ib56c254436f32fa4ed54b150715e180d329bf800
Reviewed-on: http://review.couchbase.org/88514
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>