Terminate testapp if we've got an error in SetUpTestCase
The problem is that even if this test fails we'd be running
the test case, and we don't start by asserting that there
isn't an error.

If the problem is that we failed to start the memcached
server, all test cases will print out errors reading from
the file etc.

Change-Id: I5659abafd0dd986ef2cc91d8d9bd3a65191aab5a
Reviewed-on: http://review.couchbase.org/88602
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>