MB-27790 CouchRocks: Enable RocksDB Partitioned Index and Filter
In a DGM scenario, a CMD_GET is likely to perform a bg_fetch operation
to retrieve a document.
RocksDB uses a Bloom Filter and an Index per each SST file to
help in locating data. In the ideal situation, 100% of Filter and Index
blocks are stored in the Block Cache (BC). In that case, RocksDB can load
data blocks from disk with a single read operation. But, if the ratio of
cached Filter and Index blocks is below 100%, then we can a cache-miss.
E.g., on a index-miss RocksDB needs to evict a block from the BC, search
the underlying LSM tree of SST files for the index block wanted and load
it into the BC. Also, swapping index blocks in and out as an impact on
disk utilization.
Thus, it is very important to keep the Index/Filter blocks hit-ratio as
close as possible to 100%.
With this change we use the RocksDB Partitioned Index and Filter blocks
as baseline configuration to optimize the Block Cache utilization.

Change-Id: Icd05fe1efd8a604fa5fd72b7f9a90d5d99ea971b
Reviewed-on: http://review.couchbase.org/88623
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>