Bump the timeout value for portnumber file to be created
I've seen that running under valgrind startup of the processes
might be slow, and getting even worse if the machine is under
load.

Instead of having a "false timeout" just because the
server is slow, lets set the deadline to a high value so that
if we hit it we have a real problem and not just a loaded
server (rebuilding all of the source one more time is just
putting more load on the servers).

Change-Id: I323d1298b4630aa479177394fa04814035398ef3
Reviewed-on: http://review.couchbase.org/88749
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>