MB-27769: BufferLog: add range checks to bytesOutstanding & acked
Change the types of ackedBytes to be Monotonic<> - it should never go
backwards.

Similary for bytesOutstanding - change it to NonNegativeCounter. Note
that bytesOutstanding was previously named bytesSent; however that's
somewhat misleading as it impled (to me at least) a total amount sent,
not the current anount outstanding.

Check that we don't try to acknowledge more bytes than are outstanding
(in release_UNLOCKED) - throw a logic_error if that is the case.

Change-Id: I168ff9f5ab93bf5c2d5dcae0224cfb1a19895111
Reviewed-on: http://review.couchbase.org/88768
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>