MB-27769: Add 'last_read_seqno_unsnapshotted' to stream stats
'last_read_seqno_unsnapshotted' is the last sequence number queued
from memory for a given stream, but is yet to be put in a snapshot
(to be sent out as part of DCP streaming).

With collection filters, some items may not be sent out as part
of a snapshot, but neverthless will be read on the stream and will
update the 'last_read_seqno_unsnapshotted'. Hence
'last_read_seqno_unsnapshotted' and 'last_read_seqno' may not be
same always and it helps to add the stat for debug purposes.

Change-Id: Ie9c0c95ef8a2b384b95f715b50594b58b1ca4973
Reviewed-on: http://review.couchbase.org/88805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>