Bump test timeouts
We've seen too many test timeouts lately. We have been adding
a lot of new tests to the test suites (by running the tests
with and without JSON and Compression enabled) and hence
the tests takes longer.

It is not the normal CV's job to track perf regressions in
the test suite (we've got a special CV job running perf tests)
so we don't want all of the other cv jobs failing due to
timeouts on the shared machines which may be overloaded

Change-Id: I46b0f80e81da61405184251d5348bcbadb1a4256
Reviewed-on: http://review.couchbase.org/88859
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>