MB-22010: Create ItemFreqDecayerTask
Creates an ItemFreqDecayerTask on the initialization of a KVBucket.
It is used to ensure that the frequency counters of items stored in the
hash table do not all become saturated.  When the task runs it will then
snooze for int max and will only be woken up when the frequency counter
of an item in the hash table becomes saturated.

Change-Id: I9ae51dfa6717c6349e43ddb08dde0759e43ca16b
Reviewed-on: http://review.couchbase.org/88761
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>