MB-27199: Address TSAN issues with ephemeral backfill
The memory backfill will read much of a StoredValue when it does
StoredValue::toItem. All of the StoredValue members are generally
updated under the HashBucketLock, so obtain the same for correct
access.

The linked-list code (range read etc...) often reads the bySeqno of
an entry. In general StoredValues are either out of the hashtable and
won't have their bySeqno changed by a frontend op and are protected
by the range-read lock, for elements in the hashtable, the hash-bucket
lock provides safe access for updates. However TSAN doesn't detect
this and sees someone writing the bySeqno with a hash-bucket lock and
someone reading it without the hash-bucket lock, hence the change to
StoredValue making it a RelaxedAtomic

dcp_test is updated so that a backfill test exasperates the original
issue in the MB.

Change-Id: Iadded56466b3ee92c075a3429936fcd578905b49
Reviewed-on: http://review.couchbase.org/88936
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>