Refactor: Inline cb::logtags::tagUserData
The method is so small and easy that we can just as well
keep it in the header (and as an extra bonus we can use it
directly from the other modules (like audit) without having
to add a new dependency to the utility library.

This patch also replaces the log formatting which earlier
manually inserted the <> and </> elements to use the
tag method. It was not called in the "hot path" of the
system, so it's better to have the code be clean and
use the common formatting method instead of shaving
off some us to format a debug log entry ;)

Change-Id: I6ddc41b5082ac691e0ff749f326b6032ec374364
Reviewed-on: http://review.couchbase.org/88990
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>