Expose the 'high_pri_pool_ratio' RocksDB option
In a previous patch (MB-27790) we set the
'cache_index_and_filter_blocks_with_high_priority=true' as part of the
baseline configuration to enable Partitioned Index and Filters. But, the
index/filter high-pri pool is not enabled if the 'high_pri_pool_ratio'
value is 0.0.

When the high priority pool for index and filter is enabled,
the LRU-list in LRU cache will be split into two parts, one for high-pri
blocks and one for low-pri blocks.
Data blocks will be inserted to the head of low-pri pool. Index and
filter blocks will be inserted to the head of high-pri pool. If the
total usage in the high-pri pool exceed 'capacity * high_pri_pool_ratio',
then the block at the tail of high-pri pool will overflow to the head of
the low-pri pool, after which it will compete against data blocks to stay
in cache. Eviction will start from the tail of the low-pri pool.

On DGM scenarios we usually want to fill the BlockCache with (possibly)
100% of index and filter blocks, so 'high_pri_pool_ratio=1.0' and no
low-pri space for data blocks. But, if the 100% of index and filter
blocks does not fill the BlockCache, then we want to use the available
space for data blocks.
For now we expose the new 'rocksdb_block_cache_high_pri_pool_ratio'
param to configuration, so that we can experiment different settings.
We set 0.9 as default value for the new param (i.e., 90% of the
BlockCache allocated to the high-pri pool for index/filter blocks).

Change-Id: Id06cd0a77f70748fd2c8bf6b3f5c3267b6c5d916
Reviewed-on: http://review.couchbase.org/88997
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>