MB-27839: Audit filtering using the full userid
Currently audit filtering is performed using only the user component
of a userid.  However ns_server allows ldap users to have the same id as
internal users. So  {"source" : "external", "user" : "vova"} is not the
same as {"source":"local","user":"vova"}.

Therefore we need to change the filtering to compare the full userid
and not just the "user" component.

The configure parameter "disabled_users" is also updated to
"updated_userids" to reflect that it now contains full userids.

Change-Id: Ibd7d42684573e7e8d811ecd95b007a0b7a570c47
Reviewed-on: http://review.couchbase.org/88867
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>