MB-22010: Implement statistical counter eviction policy
Implements the eviction policy for the statistical counter.  The policy
is describes as follows:

The algorithm first calculates the percentage of total memory that
we need to recover to get below the low water mark.  This percentage
is used as an initial estimate of how many items to evict from the hash
tables.

The algorithm visits each vbucket in turn.  For each vbucket we visit
every item in the hash table.  For each item we:
- Add its frequency count to a histogram.
- Use the estimate percentage to retrieve a frequency threshold from the
  histogram (Note: as an opitmisation we do not update the estimate
  on every item that is visited).
- Get the frequency count of the current item and if it below or equal
  to the threshold we evict it from the hash table.

It is possible that the percentage estimate of items to evict is not
sufficient to bring the total memory memory usage below the low water
mark.  Therefore we have an evictionMultiplier that is used to increase
the estimate if additional pass of the algorithm are required.

The multipler is increased each time an additional pass of the algorithm
is required, therefore ensuring that we evict sufficient items to drop
below the low water mark.

Change-Id: I06c5249424ab6849a9c946058e67982b12744101
Reviewed-on: http://review.couchbase.org/88927
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>