MB-27844: Audit Daemon - Add event_states object
Add an optional event_states object to version 2 of the configuration to
allow events to be dynamically enabled or disabled.

The format of the object is a map of {event_id, state} e.g.

"event_states" : {"1234" : "enabled",
                  "5678" : "disabled"}

If the event_states is not defined the configuration will still be
accepted.

Given the introduction of the event_states object, the patch also
depreciates the "disabled" list in version 2 (as it does not allow
events to be disabled).

If the "disabled" list is defined in version 2 it will be ignored.

Change-Id: I7ecf7b0cbf3807677bda31623409bb710940bfd0
Reviewed-on: http://review.couchbase.org/89168
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>