Remove create-lock in the logger
The lock is pretty useless as we don't support dynamic changes
to the logger.

If we want to use a lock for this we would need to improve the
get() method so that it'll tries to aqcuire the lock if
it thinks it needs to create the blackhole logger and
rerun the check with the lock being held to avoid racing with
other threads creating the file logger.

This is just adding complexity to the code, as the "normal"
use of the logger in production is:

    Create console logger
    do stuff
    read configuration
    replace the console logger with the file logger
    do stuff
    shut down the file logger
    exit

The blackhole logger is only used by our unit tests.

Change-Id: If11af8685f96939a7064aa9d516340997745a9f6
Reviewed-on: http://review.couchbase.org/89211
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>