MB-28082: Audit Daemon: Support filtering_permitted option
With version 2 of the audit daemon configuration a filtering_permitted
boolean attribute can be added to an event description.

If set to true it means that event can be filtered out and therefore
may not appear in the audit log.  If set to false, the event is not
permitted to be filtered out due to security reasons.

The filtering_permitted attribute is set for all auditd and memcached
event descriptions.

Re-applying, as the issue was ns_server did not support stating version
2 in audit_events.json file.  This re-apply also contains an additional
change from the original patch where we set the version to 2 in the
auditd and memcached descriptor files.

Change-Id: Ie025c64c2ca4b61b34ea64e2eba1779c04cc0580
Reviewed-on: http://review.couchbase.org/89406
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>