MB-28082: Audit Daemon: Support filtering_permitted option
With version 2 of the audit daemon configuration a filtering_permitted
boolean attribute can be added to an event description.

If set to true it means that event can be filtered out and therefore
may not appear in the audit log.  If set to false, the event is not
permitted to be filtered out due to security reasons.

The filtering_permitted attribute is set for all auditd and memcached
event descriptions.

Change-Id: I1f8df48b4fc6f8c532ea3e83883e48192eb7b67f
Reviewed-on: http://review.couchbase.org/89277
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>