CMake: speed up incremntal build via CMAKE_LINK_DEPENDS_NO_SHARED
Speed up incremental builds by not depending on linked shared library
files, only the library API.

Modification to shared library *files* will not be sufficient to
re-link targets which depend on them; however changes to the shared
library interface (i.e. headers) will still cause re-linking.

Example: memcached.exe dynamically links to libplatform.so; and
depends on the various <platform/xxx> header files. With this setting,
modifications to the implemenation of libplatform.so which don't
change it's headers will not cause memcached.exe to be re-linked.

Change-Id: Ia47750552a22405c70a168d4288b079f7d6816ac
Reviewed-on: http://review.couchbase.org/89319
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>