Simplify logger tests
Increase the file rotation size to 20MB for the "basic" tests
to avoid having to deal with multiple files. Create a separate
test batch for all of the tests which expects file rotation
with a 2k file rotation threshold.

Disable time-based flushing by setting sleeptime to 0

Dump the log contents if the deduplication tests fail (to ease
debugging from the CV machines)

Change-Id: Iec7f9ae2e0aabacf60008f140e71ee967cb3904c
Reviewed-on: http://review.couchbase.org/89320
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>