MB-26021 [5/6]: Set max_checkpoints=100 & chk_max_items=10000
Modify the checkpoint configuration to permit more granular
checkpoints: increase the maximum number of items in a checkpoint
before we can create a new checkpoint from 500 to 10,000 items; and
increase the maximum number of checkpoints from 2 to 10.

Previously max_ckpt=2 meant that you would have at most 1 closed
checkpoint; and 1 open checkpoint. If there was a large DWQ (e.g. 100,000
items) then the majority of items would be inside a single large, open
checkpoint:

   CKPT[0]:     500 items (max_items=500)
   CKPT[1]:  99,500 items

As we have to flush complete checkpoints for consistency, this forces
us to flush 99,500 items in a single checkpoint.

With this change, we allow a larger number of checkpoints, and
additionlly each checkpoint can hold more items before it is
closed. The same 100K items should now look like:

    CKPT[0]:  10,000 items (max_items=10,000)
    CKPT[1]:  10,000 items
    ...
    CKPT[9]:  10,000 items

This opens up the possibility[*] to flush in smaller chunks - we can
now flush just 10,000 items at once.

Additionally this allows us to close checkpoints at a smaller
granularity - previously if there was a DCP cursor inside CKPT[1]
(99,500 items) then until that cursor moves to the end of that
checkpoint we cannot release any of the memory associated with the
checkpoint. with a larger number of checkpoints we can essentially
free in 10K chunks.

[*] Note: It isn't yet possible to flush less than all outstanding
items - as flushVB() always fetches *all* checkpoint items for the
persistence cursor; however the next patch in the sequence will add
support for this.

Change-Id: I794a3e4b7d65bf69429f3517645080f15d0b4f65
Reviewed-on: http://review.couchbase.org/89586
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>