MB-26021 [3/6]: Rename flusher{_backfill_batch_limit -> _batch_split_trigger}
The splitting of flusher commits will be extended to include items
from checkpoints where possible - if there are multiple checkpoints
and flushing the first checkpoint would exceed the batch limit then we
will split the commit into multiple. However, checkpoints are
indivisible - we need to flush a complete checkpoint to maintain
consistency. As such, any limit we specify will be approximate; as we
cannot know exactly how many items will be in each batch.

Rename flusher_backfill_batch_limit to flusher_batch_split_trigger and
update documentation, to better reflect the new semantics of the
setting.

Change-Id: Ic450477cb719d9a380f0f0eae328194a0f6ab7ef
Reviewed-on: http://review.couchbase.org/89343
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>