Fixed log message from DCP Open
When moving to spdlogger there was a bug with missing escape of
the surrounding brace characters for the flags.

Given that DCP open isn't in the performance critical path of
the server let's format the string as a proper JSON array as
well. Ex:

INFO 40: DCP connection opened successfully. flags:["PRODUCER","INCLUDE_XATTRS"]

Change-Id: Id92c02d88818a8ec9c1a514fb5a28626af7689f7
Reviewed-on: http://review.couchbase.org/89656
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>