Use unique_ptr for Configuration::addValueChangedListener
addValueChangedListener() takes ownership of the passed-in listener
object, but only if the specified key exists in the config. If not,
then it results in a memory-leak as the listener object is passed as a
raw pointer.

Fix this problem (and in general make resource ownership more
explicit) by passing by a unique_ptr<>.

Change-Id: Id2007cee95d1ea0ee46abd420e5ab693ecd1492e
Reviewed-on: http://review.couchbase.org/89664
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>