logger: move logger read to after initialization
The logger object is looked up before we parse the memcached config
file, and hence before we have access to a file-based logger - we only
have the console (babysitter.log). This means that early memcached
status messages are not written to memcached.log; only to the
babysitter. This is a recent regression due to the spdlogger
refactoring.

Given we don't actually use the logger object until after we've parsed
memcached settings, move it to after this parsing the settings, so it
is always file-based. As a result we restore the missing log messages
to memcached.log.

Change-Id: I9b6647b010f5fb8ded78acf067d6e12897c03c9a
Reviewed-on: http://review.couchbase.org/89671
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>