MB-21916: Update get/getWithHeader to work with namespaced keys
The rollback tests failed when run with collections_prototype and
the issue was found to be the get/getWithHeader code not doing
any namespacing, so a get would fail when it should have found a key.

Simply switch the function from taking a DocKey& to a StoredDocKey& and
then perform a namespaced lookup based if KVStore configuration requires
it.

Change-Id: I18816697962695fa354152b97472247097275b37
Reviewed-on: http://review.couchbase.org/89795
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>