MB-28047: Bias the eviction histogram for items that cannot be evicted
If the storedValue is eligible for eviction then add its frequency
counter value to the frequency histogram, otherwise add the maximum
(255) to indicate that the storedValue cannot be evicted.

By adding the maximum value for each storedValue that cannot be evicted
we ensure that the histogram is biased correctly so that we get a
frequency threshold that will remove the correct number of storedValue
items.

Change-Id: I2c3546bf59f9862f88e6db9ff974d6ddaed9c2e3
Reviewed-on: http://review.couchbase.org/89803
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>