MB-27199: Fix infinite loop in StreamTest.BackfillOnly/ephemeral
The while loop checks seqno != numItems, the GAT loop is racing and
pushing the seqno up, allowing for the test to sometimes never see
when seqno == numItems, allowing for the test to hang.

Adjust the test so that the GATs complete before reading the stream
state and also make the while loop safe considering that the seqno
can now change and be unpredictable based on when the backfill and
GAT loop interacted.

Change-Id: Ia6437ff1b7d83ebdfd80482459d0f915aaec5b30
Reviewed-on: http://review.couchbase.org/89882
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>