MB-25822: Do not lock on 'releaseLock' when the ConnNofierCallback runs
In MB-25822 we see a DCPConsumer not receiving any message from the
DCPProducer for more than 360 seconds and consequently closing the
connection. Even in the case of a paused connection, a DCPProducer
should send a NOOP message to the consumer every second; the
ConnManager task is responsible for that.

On the producer logs we see that at the time of the disconnection
the ConnManager task shows very slow runtime (> 360 seconds). That
is why a NOOP has not been sent.
The cause of the very slow runtime for the ConnManager task is not clear
yet. But, on the producer logs we see that also the ConnNotifierCallback
task shows the same very long runtime (> 360 seconds).

Looking at the code executed by the ConnManager and the
ConnNotifierCallback we see that both tasks acquire a 'releaseLock'.
ConnManager locks before executing 'notifyIOComplete' on the connection,
so there could be the possibility that one slow task
(ConnNotifierCallback) is slowing down the other (ConnManager), avoiding
the ConnManager to send a NOOP message to the consumer.

ConnNotifierCallback executes 'ConnMap::notifyAllPausedConnections()',
which calls 'notifyIoComplete()' for all the connections in a
'pendingNotifications' queue of owning pointers (shared_ptr). So, at
that point we do not need any guard to access those connections.

Thus, with this change we remove the lock on 'releaseLock' from
'ConnMap::notifyAllPausedConnections()'. Doing so, we essentially remove
any possible resource contention between the ConnManager and the
ConnNotifierCallback tasks.

Change-Id: I6a7c3c971576489bb04ac24e7f475ba9ae140e43
Reviewed-on: http://review.couchbase.org/90041
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>