MB-25822: Log datetime of long schedules for NON_IO tasks
It would be useful to have the exact datetime of long unexpected
schedule for debugging, the same as we have for long runtimes.
Compared to runtime, deciding for a sensible expected schedule
time is more difficult, mainly for 2 reasons:
- the schedule time depends on things "external" to the task itself
  (e.g., the number of tasks in the same priority-group)
- the schedule time depends on the runtime of the previous run of the
  same task; so, for IO tasks (Read/Write/AuxIO) is even more difficult
  to predict
Thus, with this chagnge we log unexpected long schedule time only for
NON_IO tasks. We pick 1 seconds as schedule overhead limit and log only
schedule overheads that exceed that limit.

Change-Id: I68b921b783c7bcd9644565e7e9a14f2497763ce0
Reviewed-on: http://review.couchbase.org/90047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>