MB-28327: Initialize all storedValue with correct frequency count
Changed the constructor of an Item from taking a NRU value, which
almost always uses the default value, to instead be initialized directly
to INITIAL_NRU_VALUE.  Then in the two places that we used to initialize
with a different value, we explictly set the NRU value.

Wth the above change in place, we initialise the FrequencyCounter in a
similar manner to the NRU value.  This ensures that storedValue is
initialized with the correct frequency count.

The frequency counter should be initialised to
ItemEviction::initialFreqCounter.

Change-Id: I34096cf72290298e8bf8bc84665778b3d7208690
Reviewed-on: http://review.couchbase.org/90227
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>