MB-28428: Inflate snappy document before pruning xattrs
Delete with meta of an xattr document, needs to examine any existing
document to see if the system xattrs shall remain in-tact.

If the xattr document being examined is also snappy compressed then
we must inflate that document.

In this change once we've inflated the deleted item, we don't deflate
the remaining value, it doesn't seem worth the effort given that only
system-xattrs remain.

Test added which would crash before the vbucket.cc changes.

Change-Id: I0b1fd6af7176b960fcaaecc6f3811e70bd6285d1
Reviewed-on: http://review.couchbase.org/90244
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>