MB-28453: [2/n] Switch xattrs from byte_buffer to char_buffer
Long-term we wish to remove the use of byte_buffer as we frequently
end up casting it to char (and the nonstop reinterpret_cast is
tedious). This commit just switches xattr code from using a
byte_buffer to char_buffer, making it easier to add decompression
support to xattr::Blob which is required to cleanly address the MB
this commit is linked with.

A couchstore change is linked to this as dbdump.cc requires an update
to keep working with the Blob changes.

Change-Id: I197723539ea1f7c799ccf3dcddfa1757531e36ba
Reviewed-on: http://review.couchbase.org/90363
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>