MB-27489: Allow for '*' as a bucket name
ns_server doesn't synchronize bucket deletion with updating
the RBAC files, so it may remove access to the bucket before
all of its connections to the bucket is correctly terminated.

Given that ns_server should have access to all buckets we
can add a wildcard entry which match all bucket names. Given
that "*" isn't a valid bucket name we can use that to
represent all bucket.

Change-Id: I325f7cba6fb72f38ecbf94640b4102322a5f80fa
Reviewed-on: http://review.couchbase.org/90407
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>